requirements.txt
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-stubs==1.5.0
factory-boy==3.0.1
graphene-django==2.13.0
graphene-file-upload==1.2.2
petl==1.6.4
PyYAML==5.3.1
PyJWT==1.7.1
psycopg2==2.9.1
requests==2.27.1
beautifulsoup4==4.10.0
validators==0.18.2
cfscrape==2.1.1
html-to-json==2.0.0
dicttoxml==1.7.4
werkzeug==2.1.1